... '9f374b7b17d1cb6d5921c62fcb7cfc61e2e8e5a0' into HEAD

+++ include/modem/location.h
@@ -3,14 +3,6 @@
  *
  * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
  */
-
-/**
- * @file location.h
- * @brief Public APIs for the Location library.
- * @defgroup location Location library
- * @{
- */
-
 #ifndef LOCATION_H_
 #define LOCATION_H_
 
@@ -27,6 +19,13 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
+
+/**
+ * @file location.h
+ * @brief Public APIs for the Location library.
+ * @defgroup location Location library
+ * @{
+ */
 
 /** Location method. */
 enum location_method {
@@ -155,8 +154,8 @@ struct location_event_data {
 /** GNSS configuration. */
 struct location_gnss_config {
 	/**
-	 * @brief Timeout (in seconds), meaning how long GNSS is allowed to run when trying to
-	 * acquire a fix. Zero means that the timer is disabled, meaning that GNSS will
+	 * @brief Timeout (in milliseconds), meaning how long GNSS is allowed to run when trying to
+	 * acquire a fix. SYS_FOREVER_MS means that the timer is disabled, meaning that GNSS will
 	 * continue to search until it gets a fix or the application calls cancel.
 	 *
 	 * @details Note that this is not real time as experienced by the user.
@@ -166,7 +165,7 @@ struct location_gnss_config {
 	 * and A-GPS is disabled, library waits until modem enters PSM before starting GNSS,
 	 * thus maximizing uninterrupted operating window and minimizing power consumption.
 	 */
-	uint16_t timeout;
+	int32_t timeout;
 
 	/** @brief Desired accuracy level.
 	 *
@@ -203,10 +202,10 @@ struct location_gnss_config {
 /** LTE cellular positioning configuration. */
 struct location_cellular_config {
 	/**
-	 * @brief Timeout (in seconds) of how long the cellular positioning procedure can take.
-	 * Zero means that the timer is disabled.
+	 * @brief Timeout (in milliseconds) of how long the cellular positioning procedure can take.
+	 * SYS_FOREVER_MS means that the timer is disabled.
 	 */
-	uint16_t timeout;
+	int32_t timeout;
 
 	/** Used cellular positioning service. */
 	enum location_service service;
@@ -215,10 +214,10 @@ struct location_cellular_config {
 /** Wi-Fi positioning configuration. */
 struct location_wifi_config {
 	/**
-	 * @brief Timeout (in seconds) of how long the Wi-Fi positioning procedure can take.
-	 * Zero means that the timer is disabled.
+	 * @brief Timeout (in milliseconds) of how long the Wi-Fi positioning procedure can take.
+	 * SYS_FOREVER_MS means that the timer is disabled.
 	 */
-	uint16_t timeout;
+	int32_t timeout;
 
 	/** Used Wi-Fi positioning service. */
 	enum location_service service;
